LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/event-error-expected.txt

Blocked access to external URL http://example.nonexist/
CONSOLE MESSAGE: line 22: XMLHttpRequest cannot load http://example.nonexist/ due to access control checks.

PASS XMLHttpRequest Test: event - error 
